Moving kernel_approximation to the top level
