+++ sklearn/feature_extraction/__init__.py
@@ -6,4 +6,3 @@ images.
 
 from .image import img_to_graph, grid_to_graph
 from . import text
-from .kernel_approximation import RBFSampler, SkewedChi2Sampler, AdditiveChi2Sampler
